+++ src/Graphics/TextureResource/Abstract.hpp
@@ -199,10 +199,10 @@ namespace EmEn::Graphics::TextureResource
 			/**
 			 * @brief Returns the index of the frame at specific time.
 			 * @param sceneTime The current scene time in milliseconds.
-			 * @return size_t
+			 * @return uint32_t
 			 */
 			[[nodiscard]]
-			virtual size_t frameIndexAt (uint32_t sceneTime) const noexcept = 0;
+			virtual uint32_t frameIndexAt (uint32_t sceneTime) const noexcept = 0;
 
 			/**
 			 * @brief Returns the image of the texture.
